module/cmn_cyprus: fix incorrect format specifier for clang
The Clang compiler considers uint32_t data type as 'unsigned int'
whereas the GCC considers it as 'long unsigned int'. A potential
solution is to use PRIx format specifiers provided in the inttypes
header but it is not compatible with aarch64 host machines.

Hence, as a workaround, cast the uint32_t variables as unsigned int data
type when printing as the concerned variables are guaranteed to be under
16 bits.

+++ module/cmn_cyprus/src/cmn_cyprus_rnsam_setup.c
@@ -1,6 +1,6 @@
 /*
  * Arm SCP/MCP Software
- * Copyright (c) 2023, Arm Limited and Contributors. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited and Contributors. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -98,7 +98,8 @@ static int program_io_region(
 
     if (region_idx >= RNSAM_NON_HASH_MEM_REGION_COUNT) {
         FWK_LOG_ERR(
-            MOD_NAME "Error! Invalid non-hashed region %lu", region_idx);
+            MOD_NAME "Error! Invalid non-hashed region %u",
+            (unsigned int)region_idx);
         FWK_LOG_ERR(
             MOD_NAME "Max non-hashed region supported is %u",
             RNSAM_NON_HASH_MEM_REGION_COUNT);
@@ -228,7 +229,8 @@ static int program_scg_region(
                      ->rnsam_scg_config;
 
     if (scg_idx >= MAX_SCG_COUNT) {
-        FWK_LOG_ERR(MOD_NAME "Error! Invalid SCG region %lu", scg_idx);
+        FWK_LOG_ERR(
+            MOD_NAME "Error! Invalid SCG region %u", (unsigned int)scg_idx);
         FWK_LOG_ERR(MOD_NAME "Max SCG region supported is %u ", MAX_SCG_COUNT);
         return FWK_E_DATA;
     }
